add some (long) casting to make sure the compiler doesn't do weird things with byte counts above 16GB

+++ cbench/opensource/HPLlinpack/hpl/testing/ptest/HPL_pdtest.c
@@ -160,7 +160,7 @@ void HPL_pdtest
 /*
  * Allocate dynamic memory
  */
-   vptr = (void*)malloc( (ALGO->align + (mat.ld+1)*(mat.nq))*sizeof(double) );
+   vptr = (void*)malloc((long)(ALGO->align + (long)(mat.ld+1)*(mat.nq))*sizeof(double) );
    info[0] = (vptr == NULL); info[1] = myrow; info[2] = mycol;
    (void) HPL_all_reduce( (void *)(info), 3, HPL_INT, HPL_max,
                           GRID->all_comm );
